Move input error handling from upload to parsing stage
Previously, some errors, e.g. "no files for format" were only checked
during the upload. This means that if an error occurs in this stage,
the upload session is terminated, but the session is not closed.
I have now moved the input validation to the parsing stage, so that
all errors should already be found before the upload session is opened.